SMU.cpp: Modify the scanning process to align with changes in libsmu.
We now identify devices using a combination of
USB bus_id and USB address. This way we can detect if a running device
was disconnected (without interfering with the signal) or a new one was
connected.
After detecting that the device is already connected, we don't need to
delete the object, because it's the same one that's running a capture.

Signed-off-by: Alexandra Trifan <Alexandra.Trifan@analog.com>